Tested Calculator and Server
Tested the celestrial coordinate calculator with the telescope manager
and tested the Server running with the scheduler by implementing a dummy
calculation to slew the goto back to stellarium as done in their example

+++ Software/Src/StellariumServer/Socket.hpp
@@ -87,7 +87,7 @@ class Socket
      * @param DecInt int32_t version of the Declination
      * @param Status int32_t version of the 
      */
-        virtual void SendPosition( uint16_t RAInt, int32_t DecInt, int32_t Status) {}
+        virtual void SendPosition( uint32_t RAInt, int32_t DecInt, int32_t Status) {}
         
     protected:
     /** Constructor
